+++ libsemanage/src/semanage_store.c
@@ -137,10 +137,10 @@ typedef struct semanage_file_context_node {
 	char *path;
 	char *file_type;
 	char *context;
-	int path_len;
-	int effective_len;
-	int type_len;
-	int context_len;
+	size_t path_len;
+	size_t effective_len;
+	size_t type_len;
+	size_t context_len;
 	int meta;		/* position of first meta char in path, -1 if none */
 	struct semanage_file_context_node *next;
 } semanage_file_context_node_t;
@@ -514,7 +514,7 @@ const char *semanage_final_path(enum semanage_final_defs store,
 char *semanage_conf_path(void)
 {
 	char *semanage_conf = NULL;
-	int len;
+	size_t len;
 	struct stat sb;
 
 	len = strlen(semanage_root()) + strlen(selinux_path()) + strlen(SEMANAGE_CONF_FILE);
@@ -2893,7 +2893,7 @@ int semanage_nc_sort(semanage_handle_t * sh, const char *buf, size_t buf_len,
 
 	/* parsing bits */
 	const char *priority_names[] = NC_SORT_NAMES;
-	const int priority_names_len[] = NC_SORT_NAMES_LEN;
+	const size_t priority_names_len[] = NC_SORT_NAMES_LEN;
 	size_t line_len, buf_remainder, i, offset;
 	const char *line_buf, *line_end;
